+++ example/classes.dart
@@ -2,7 +2,7 @@ import 'package:dungeon_world_data/dw_data.dart';
 
 void main() {
   // Get bard class
-  var bard = dungeonWorld.classes.firstWhere((k) => k.key == 'bard');
+  var bard = DWEntity.getByKey('bard', dungeonWorld.classes);
   // Get all bard advanced moves
 
   var moves = bard.advancedMoves1;