+++ example/moves.dart
@@ -1,8 +1,8 @@
 import 'package:dungeon_world_data/dw_data.dart';
 
 void main() {
-  var bard = dungeonWorld.classes.firstWhere((k) => k.key == 'bard');
-  var thief = dungeonWorld.classes.firstWhere((k) => k.key == 'thief');
+  var bard = DWEntity.getByKey('bard', dungeonWorld.classes);
+  var thief = DWEntity.getByKey('thief', dungeonWorld.classes);
 
   // Flattened move lists
   var moves1 = bard.advancedMoves1;