+++ lib/dice.dart
@@ -2,16 +2,24 @@ import 'dart:math';
 import 'package:quiver/core.dart';
 
 class Dice {
+  // Amount of dice in set
   num amount;
+
+  // No. of sides for the die
   num sides;
+
+  // Modifier value for dice, can be added to `DiceResult` value
+  num modifier;
+
+  // Last result rolled by this die
   DiceResult lastResult;
 
-  /// Simple dice, with sides and die count.
+  /// Simple dice, with sides, die count and modifier.
   /// You can multiply, add or subtract Dice objects to change the amount of rolls (notice dice must
   /// be with the same amount of sides).
   ///
   /// To roll multiple dice at once, you may use the static `Dice.roll(<Dice>[...])`.
-  Dice(this.sides, [this.amount = 1]);
+  Dice(this.sides, [this.amount = 1, this.modifier = 0]);
 
   static Dice d4 = Dice(4);
   static Dice d6 = Dice(6);
@@ -19,6 +27,12 @@ class Dice {
   static Dice d10 = Dice(10);
   static Dice d12 = Dice(12);
   static Dice d20 = Dice(20);
+
+  Dice copyWith({int sides, int amount, int modifier}) => Dice(
+        sides ?? this.sides,
+        amount ?? this.amount,
+        modifier ?? this.modifier,
+      );
 
   /// Parse strings such as `d6`, or `2d20` into a `Dice` object.
   static Dice parse(String str) {
@@ -37,7 +51,7 @@ class Dice {
       if (obj.sides != sides) {
         throw ("Can't multiply different sided die!");
       }
-      return Dice(sides, obj.amount * amount);
+      return Dice(sides, obj.amount * amount, modifier);
     }
 
     if (obj is num) {
@@ -52,11 +66,11 @@ class Dice {
       if (obj.sides != sides) {
         throw ("Can't divide different sided die!");
       }
-      return Dice(sides, obj.amount / amount);
+      return Dice(sides, obj.amount / amount, modifier);
     }
 
     if (obj is num) {
-      return Dice(sides, amount ~/ obj);
+      return Dice(sides, amount ~/ obj, modifier);
     }
 
     return this;
@@ -67,11 +81,11 @@ class Dice {
       if (obj.sides != sides) {
         throw ("Can't add different sided die!");
       }
-      return Dice(sides, obj.amount + amount);
+      return Dice(sides, obj.amount + amount, modifier);
     }
 
     if (obj is num) {
-      return Dice(sides, amount + obj.toInt());
+      return Dice(sides, amount + obj.toInt(), modifier);
     }
 
     return this;
@@ -82,11 +96,11 @@ class Dice {
       if (obj.sides != sides) {
         throw ("Can't subtract different sided die!");
       }
-      return Dice(sides, obj.amount - amount);
+      return Dice(sides, obj.amount - amount, modifier);
     }
 
     if (obj is num) {
-      return Dice(sides, amount - obj.toInt());
+      return Dice(sides, amount - obj.toInt(), modifier);
     }
 
     return this;
@@ -95,14 +109,16 @@ class Dice {
   @override
   bool operator ==(obj) {
     if (obj is Dice) {
-      return amount == obj.amount && sides == obj.sides;
+      return amount == obj.amount &&
+          sides == obj.sides &&
+          modifier == obj.modifier;
     }
 
     return obj.toString() == toString();
   }
 
   @override
-  int get hashCode => hash2(amount, sides);
+  int get hashCode => hash3(amount, sides, modifier);
 
   /// Rolls the dice and returns the `DiceResult`.
   DiceResult getRoll() {
@@ -125,6 +141,8 @@ class Dice {
 
     return results;
   }
+
+  String get modRepr => (modifier > 0 ? '+' : '') + modifier.toString();
 }
 
 class DiceResult {
@@ -138,18 +156,27 @@ class DiceResult {
   DiceResult(this.dice, this.results);
 
   @override
-  String toString() => '$dice${hit20 ? '*' : ''} => $total';
-  String get detailed =>
-      '$dice${hit20 ? '*' : ''} => $total\n  $mappedResults\n  ${hit20 ? "Die no. ${hit20At} hit 20" : "Didn\'t hit 20"}';
+  String toString() => '$dice${didHitNaturalMax ? '*' : ''} => $total';
+
+  // More detailed version of `toString`.
+  String get toDetailedString =>
+      '$dice${didHitNaturalMax ? '*' : ''} => $total\n  $mappedResults\n  ${didHitNaturalMax ? "Die no. ${indexOfNaturalMax} hit 20" : "Didn\'t hit 20"}';
+
+  // All results layed out with their respective die.
   String get mappedResults {
     var out = <String>[];
     for (num i = 0; i < results.length; i++) {
       out.add('${i + 1}: ${results[i]}');
     }
-    return out.toString();
+    return out.toString() + (dice.modifier != 0 ? ' (${dice.modRepr})' : '');
   }
 
-  num get total => results.reduce((tot, cur) => tot + cur);
-  bool get hit20 => results.any((r) => r == 20);
-  num get hit20At => hit20 ? results.indexOf(20) : null;
+  /// Total (accumulated) value of result, including modifiers.
+  num get total => results.reduce((tot, cur) => tot + cur) + dice.modifier;
+
+  // Boolean that represents whether any of the rolled dice hit their natural max value (e.g. 20 for d20)
+  bool get didHitNaturalMax => results.any((r) => r == dice.sides);
+
+  // Get first index of the die that hit natural max value.
+  num get indexOfNaturalMax => results.indexOf(dice.sides);
 }