+++ test/dw_entity_test.dart
@@ -25,5 +25,18 @@ void main() {
       var t2 = Test1(key: '1', value: 1);
       expect(t1, equals(t2));
     });
+
+    group('Helpers', () {
+      test('getByKey', () {
+        var classes = dungeonWorld.classes;
+        var wizard = dungeonWorld.classes.firstWhere((k) => k.key == 'wizard');
+
+        var found = DWEntity.getByKey<PlayerClass>('wizard', classes);
+        expect(wizard, equals(found));
+
+        var notFound = DWEntity.getByKey<PlayerClass>("doesn't exist", classes);
+        expect(notFound, equals(null));
+      });
+    });
   });
 }