+++ test/equipment_test.dart
@@ -4,8 +4,7 @@ import 'package:dungeon_world_data/dw_data.dart';
 void main() {
   group('Equipment', () {
     test('Key', () {
-      var equipment =
-          dungeonWorld.equipment.firstWhere((k) => k.key == 'tricksy_rope');
+      var equipment = DWEntity.getByKey('tricksy_rope', dungeonWorld.equipment);
       expect(equipment.key, equals('tricksy_rope'));
     });
   });