+++ test/player_class_test.dart
@@ -4,9 +4,8 @@ import 'package:dungeon_world_data/dw_data.dart';
 void main() {
   group('PlayerClass', () {
     test('Key', () {
-      var playerClass =
-          dungeonWorld.classes.firstWhere((k) => k.key == 'paladin');
-      expect(playerClass.key, equals('paladin'));
+      var paladin = DWEntity.getByKey('paladin', dungeonWorld.classes);
+      expect(paladin.key, equals('paladin'));
     });
   });
 }